pyproject.toml
[project]
name = "downbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.2",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.8.11",
    "boto3>=1.35.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.12",
]
